Cleaning up main function
